requirements.txt
flask==3.1.0
requests
Werkzeug
pydantic==2.10.6
pytelegrambotapi==4.26.0
gunicorn
ua_parser
SQLAlchemy==2.0.38
flask-sqlalchemy==3.1.1
flask-limiter==3.3.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1